Made ReferencedSymbolFinder private
